requirements.txt
Flask==2.3.2
flask-pymongo==2.3.0
python-docx==0.8.11
pymongo==4.4.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2


